Crate solana_instruction

Types for directing the execution of Solana programs.

Every invocation of a Solana program executes a single instruction, as defined by the Instruction type. An instruction is primarily a vector of bytes, the contents of which are program-specific, and not interpreted by the Solana runtime. This allows flexibility in how programs behave, how they are controlled by client software, and what data encodings they use.

Besides the instruction data, every account a program may read or write while executing a given instruction is also included in Instruction, as AccountMeta values. The runtime uses this information to efficiently schedule execution of transactions.

Re-exports

pub use account_meta::AccountMeta;

Modules

account_meta

error

Structs

BorrowedAccountMeta

Borrowed version of AccountMeta.

BorrowedInstruction

Borrowed version of Instruction.

Instruction

A directive for a single invocation of a Solana program.

ProcessedSiblingInstruction

Use to query and convey information about the sibling instruction components when calling the sol_get_processed_sibling_instruction syscall.

Constants

TRANSACTION_LEVEL_STACK_HEIGHT